requirements.txt
Django>=5.0,<6.0
psycopg2-binary>=2.9
requests>=2.31.0
python-telegram-bot>=21.0.0

